feat: improve messages components
