+++ package.json
@@ -22,9 +22,9 @@
     "@cosmjs/ledger-amino": "^0.29.5",
     "@cosmjs/proto-signing": "^0.29.5",
     "@cosmjs/stargate": "^0.29.5",
-    "@desmoslabs/desmjs": "^4.7.3",
-    "@desmoslabs/desmjs-types": "^4.7.3",
-    "@desmoslabs/desmjs-walletconnect-v2": "^4.7.3",
+    "@desmoslabs/desmjs": "^4.7.4",
+    "@desmoslabs/desmjs-types": "^4.7.4",
+    "@desmoslabs/desmjs-walletconnect-v2": "^4.7.4",
     "@habx/apollo-multi-endpoint-link": "^2.7.0",
     "@json-rpc-tools/types": "^2.0.0-beta.12",
     "@json-rpc-tools/utils": "^2.0.0-beta.12",
