+++ src/components/Messages/MsgDelegate/MsgDelegateDetails/useStyles.ts
@@ -1,0 +1,10 @@
+import { makeStyle } from 'config/theme';
+
+const useStyles = makeStyle((_) => ({
+  validatorAddress: {
+    flexDirection: 'row',
+    flexShrink: 1,
+  },
+}));
+
+export default useStyles;
