+++ src/components/Messages/MsgDelegate/index.ts
@@ -1,0 +1,11 @@
+import { MsgDelegateEncodeObject } from '@cosmjs/stargate';
+import { MessageComponents } from 'components/Messages/BaseMessage';
+import MsgDelegateListItem from './MsgDelegateListItem';
+import MsgDelegateDetails from './MsgDelegateDetails';
+
+const MsgDelegateComponents: MessageComponents<MsgDelegateEncodeObject['value']> = {
+  details: MsgDelegateDetails,
+  listItem: MsgDelegateListItem,
+};
+
+export default MsgDelegateComponents;
