+++ src/components/Messages/MsgSend/MsgDetailsListItem/index.tsx
@@ -1,0 +1,42 @@
+import { MsgSendEncodeObject } from '@cosmjs/stargate';
+import React, { useMemo } from 'react';
+import { useTranslation } from 'react-i18next';
+import { formatCoins } from 'lib/FormatUtils';
+import BaseMessageDetails from 'components/Messages/BaseMessage/BaseMessageDetails';
+import { msgSendIcon } from 'assets/images';
+
+export type MsgSendDetailsProps = {
+  message: MsgSendEncodeObject['value'];
+};
+
+/**
+ * Displays the full details of a MsgSend.
+ */
+const MsgSendDetails = (props: MsgSendDetailsProps) => {
+  const { t } = useTranslation('sendTokens');
+  const { message } = props;
+  const convertedAmount = useMemo(() => formatCoins(message.amount), [message.amount]);
+
+  return (
+    <BaseMessageDetails
+      icon={msgSendIcon}
+      iconSubtitle={convertedAmount}
+      fields={[
+        {
+          label: t('from'),
+          value: message.fromAddress ?? '',
+        },
+        {
+          label: t('to'),
+          value: message.toAddress ?? '',
+        },
+        {
+          label: t('amount'),
+          value: convertedAmount,
+        },
+      ]}
+    />
+  );
+};
+
+export default MsgSendDetails;
