+++ src/components/Messages/MsgUnlinkChainAccount/index.tsx
@@ -1,102 +1,13 @@
 import { MsgUnlinkChainAccountEncodeObject } from '@desmoslabs/desmjs';
-import React, { useCallback, useMemo } from 'react';
-import { useTranslation } from 'react-i18next';
-import { Image, StyleSheet, View } from 'react-native';
-import Typography from 'components/Typography';
-import SupportedChains from 'config/LinkableChains';
-import { cosmosIcon, desmosIcon, disconnectIcon, msgGeneralIcon } from 'assets/images';
-import BaseMessage from '../BaseMessage';
-import useStyles from './useStyles';
+import { MessageComponents } from 'components/Messages/BaseMessage';
+import MsgUnlinkChainAccountDetails from './MsgUnlinkChainAccountDetails';
+import MsgUnlinkChainAccountListItem from './MsgUnllinkChainAccountListItem';
 
-export type DetailsProps = {
-  message: MsgUnlinkChainAccountEncodeObject['value'];
+const MsgUnlinkChainAccountComponents: MessageComponents<
+  MsgUnlinkChainAccountEncodeObject['value']
+> = {
+  details: MsgUnlinkChainAccountDetails,
+  listItem: MsgUnlinkChainAccountListItem,
 };
 
-export type ListItemProps = {
-  message: MsgUnlinkChainAccountEncodeObject['value'];
-  date: Date;
-};
-
-namespace MsgUnlinkChainAccount {
-  /**
-   * Displays the short details of a MsgUnlinkChainAccount within a list.
-   * @constructor
-   */
-  export const ListItem: React.FC<ListItemProps> = ({ message, date }) => {
-    const { t } = useTranslation();
-    const styles = useStyles();
-
-    return (
-      <BaseMessage.ListItem
-        icon={msgGeneralIcon}
-        date={date}
-        renderContent={() => (
-          <View>
-            <Typography.Body1>{t('tx type unlink chain account')}</Typography.Body1>
-            <View style={styles.target}>
-              <Typography.Caption>{message.target}</Typography.Caption>
-            </View>
-          </View>
-        )}
-      />
-    );
-  };
-
-  /**
-   * Displays the full details of a MsgUnlinkChainAccount
-   * @constructor
-   */
-  export const Details: React.FC<DetailsProps> = ({ message }) => {
-    const styles = StyleSheet.create({
-      chainLinkIcon: {
-        height: 44,
-        width: 44,
-      },
-      customIconView: {
-        alignItems: 'center',
-        display: 'flex',
-        flexDirection: 'row',
-      },
-      disconnectIcon: {
-        height: 24,
-        marginHorizontal: 20,
-        width: 24,
-      },
-    });
-
-    const { t } = useTranslation();
-    const getLinkableChainInfoByName = useCallback(
-      (chainName: string) => SupportedChains.find((chain) => chainName === chain.chainConfig.name),
-      [],
-    );
-
-    const { chainName } = message;
-    const chainIcon = useMemo(() => {
-      const chain = chainName !== undefined ? getLinkableChainInfoByName(chainName) : undefined;
-      if (chain !== undefined) {
-        return chain.icon;
-      }
-      return cosmosIcon;
-    }, [chainName, getLinkableChainInfoByName]);
-
-    return (
-      <BaseMessage.Details
-        customIconView={
-          <View style={styles.customIconView}>
-            <Image style={styles.chainLinkIcon} source={desmosIcon} />
-            <Image style={styles.disconnectIcon} source={disconnectIcon} />
-            <Image style={styles.chainLinkIcon} source={chainIcon} />
-          </View>
-        }
-        fields={[
-          {
-            label: t('unlinked account'),
-            value: message.target,
-          },
-        ]}
-      />
-    );
-  };
-}
-
-export default MsgUnlinkChainAccount;
+export default MsgUnlinkChainAccountComponents;
